feat: remove old code, add prefetch, market, currency all
