+++ apps/ledger-live-desktop/src/renderer/screens/account/AccountHeaderActions.tsx
@@ -22,7 +22,6 @@ import Tooltip from "~/renderer/components/Tooltip";
 import useTheme from "~/renderer/hooks/useTheme";
 import IconAccountSettings from "~/renderer/icons/AccountSettings";
 import IconWalletConnect from "~/renderer/icons/WalletConnect";
-import { useProviders } from "~/renderer/screens/exchange/Swap2/Form";
 import { rgba } from "~/renderer/styles/helpers";
 import { track } from "~/renderer/analytics/segment";
 import {
@@ -37,6 +36,8 @@ import { useGetSwapTrackingProperties } from "~/renderer/screens/exchange/Swap2/
 import { CryptoCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
 import { getLLDCoinFamily } from "~/renderer/families";
 import { ManageAction } from "~/renderer/families/types";
+import { getAvailableProviders } from "@ledgerhq/live-common/exchange/swap/index";
+import { useFetchCurrencyAll } from "@ledgerhq/live-common/exchange/swap/hooks/index";
 
 type RenderActionParams = {
   label: React.ReactNode;
@@ -147,6 +148,7 @@ const AccountHeaderSettingsButtonComponent = ({ account, parentAccount, openModa
 };
 const pageName = "Page Account";
 const AccountHeaderActions = ({ account, parentAccount, openModal }: Props) => {
+  const { data: currenciesAll } = useFetchCurrencyAll();
   const mainAccount = getMainAccount(account, parentAccount);
   const contrastText = useTheme().colors.palette.text.shade60;
   const swapDefaultTrack = useGetSwapTrackingProperties();
@@ -179,13 +181,10 @@ const AccountHeaderActions = ({ account, parentAccount, openModal }: Props) => {
       allSellableCryptoCurrencyIds.includes(currency.id),
     ];
   }, [rampCatalog.value, currency.id]);
-  const { providers, storedProviders, providersError } = useProviders();
 
   // don't show buttons until we know whether or not we can show swap button, otherwise possible click jacking
-  const showButtons = !!(providers || storedProviders || providersError);
-  const availableOnSwap = providers?.concat(storedProviders ?? []).some(({ pairs }) => {
-    return pairs && pairs.find(({ from, to }) => [from, to].includes(currency.id));
-  });
+  const showButtons = !!getAvailableProviders();
+  const availableOnSwap = (currenciesAll ?? []).includes(currency.id);
 
   const history = useHistory();
   const buttonSharedTrackingFields = useMemo(
