+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/Form/FormSelectors/FromRow.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect } from "react";
-import { useSelector } from "react-redux";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import {
@@ -8,13 +7,11 @@ import {
   getAccountName,
 } from "@ledgerhq/live-common/account/index";
 import Box from "~/renderer/components/Box";
-import { fromSelector } from "~/renderer/actions/swap";
 import InputCurrency from "~/renderer/components/InputCurrency";
 import { ErrorContainer } from "~/renderer/components/Input";
 import { SelectAccount } from "~/renderer/components/SelectAccount";
 import Switch from "~/renderer/components/Switch";
 import Text from "~/renderer/components/Text";
-import { shallowAccountsSelector } from "~/renderer/reducers/accounts";
 import { amountInputContainerProps, renderAccountValue, selectRowStylesMap } from "./utils";
 import { FormLabel } from "./FormLabel";
 import {
@@ -30,6 +27,7 @@ import { AccountLike } from "@ledgerhq/types-live";
 import BigNumber from "bignumber.js";
 import { TranslatedError } from "~/renderer/components/TranslatedError/TranslatedError";
 import { WarningSolidMedium } from "@ledgerhq/react-ui/assets/icons";
+import { useSwapableAccounts } from "../hooks/useSwapableAccounts";
 
 const SwapStatusContainer = styled.div<{ isError: boolean }>(
   ({ theme: { space, colors }, isError }) => `
@@ -123,7 +121,7 @@ function FromRow({
   updateSelectedRate,
 }: Props) {
   const swapDefaultTrack = useGetSwapTrackingProperties();
-  const accounts = useSelector(fromSelector)(useSelector(shallowAccountsSelector));
+  const accounts = useSwapableAccounts();
   const unit = fromAccount && getAccountUnit(fromAccount);
   const { t } = useTranslation();
   usePickDefaultAccount(accounts, fromAccount, setFromAccount);
