+++ apps/ledger-live-desktop/src/renderer/screens/market/MarketCoinScreen/index.tsx
@@ -15,7 +15,6 @@ import { useGetSwapTrackingProperties } from "~/renderer/screens/exchange/Swap2/
 import { Button } from "..";
 import MarketCoinChart from "./MarketCoinChart";
 import MarketInfo from "./MarketInfo";
-import { useProviders } from "../../exchange/Swap2/Form";
 import { getAvailableAccountsById } from "@ledgerhq/live-common/exchange/swap/utils/index";
 import { accountsSelector } from "~/renderer/reducers/accounts";
 import { openModal } from "~/renderer/actions/modals";
@@ -25,6 +24,7 @@ import { flattenAccounts } from "@ledgerhq/live-common/account/index";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import useStakeFlow from "../../stake";
 import { stakeDefaultTrack } from "~/renderer/screens/stake/constants";
+import { useFetchCurrencyAll } from "@ledgerhq/live-common/exchange/swap/hooks/index";
 
 const CryptoCurrencyIconWrapper = styled.div`
   height: 56px;
@@ -67,16 +67,8 @@ export default function MarketCoinScreen() {
   const locale = useSelector(localeSelector);
   const allAccounts = useSelector(accountsSelector);
   const flattenedAccounts = flattenAccounts(allAccounts);
-  const { providers, storedProviders } = useProviders();
   const swapDefaultTrack = useGetSwapTrackingProperties();
-
-  const swapAvailableIds = useMemo(() => {
-    return providers || storedProviders
-      ? (providers || storedProviders)!
-          .map(({ pairs }) => pairs.map(({ from, to }) => [from, to]))
-          .flat(2)
-      : [];
-  }, [providers, storedProviders]);
+  const { data: currenciesAll } = useFetchCurrencyAll();
 
   const {
     selectedCoinData: currency,
@@ -124,7 +116,7 @@ export default function MarketCoinScreen() {
 
   const availableOnBuy =
     currency && currency.ticker && onRampAvailableTickers.includes(currency.ticker?.toUpperCase());
-  const availableOnSwap = internalCurrency && swapAvailableIds.includes(internalCurrency.id);
+  const availableOnSwap = internalCurrency && (currenciesAll ?? []).includes(internalCurrency.id);
   const stakeProgramsFeatureFlag = useFeature("stakePrograms");
   const listFlag = stakeProgramsFeatureFlag?.params?.list ?? [];
   const stakeProgramsEnabled = stakeProgramsFeatureFlag?.enabled ?? false;
