+++ libs/ledger-live-common/src/exchange/swap/api/v5/__mocks__/fetchCurrencyAll.mocks.ts
@@ -1,6 +1,6 @@
-import { ResponseData } from "../fetchCurrencyAll";
+import { ResponseDataAll } from "../fetchCurrencyAll";
 
-export const fetchCurrencyAllMock: ResponseData = {
+export const fetchCurrencyAllMock: ResponseDataAll = {
   from: [
     {
       network: "ethereum",
