+++ libs/ledger-live-common/src/exchange/swap/api/v5/fetchCurrencyAll.ts
@@ -5,35 +5,34 @@ import { DEFAULT_SWAP_TIMEOUT_MS } from "../../const/timeout";
 import axios from "axios";
 import { LedgerAPI4xx } from "@ledgerhq/errors";
 import { fetchCurrencyAllMock } from "./__mocks__/fetchCurrencyAll.mocks";
+import { ResponseData as ResponseDataTo } from "./fetchCurrencyTo";
+import { ResponseData as ResponseDataFrom } from "./fetchCurrencyFrom";
+import { flattenV5CurrenciesAll } from "../../utils/flattenV5CurrenciesAll";
 
 type Props = {
   providers: Array<string>;
   additionalCoinsFlag?: boolean;
 };
 
-export type ResponseData = {
-  from: From[];
-  to: From[];
-};
-type From = {
-  network: string;
-  supportedCurrencies: string[];
+export type ResponseDataAll = {
+  from: ResponseDataFrom["currencyGroups"];
+  to: ResponseDataTo["currencyGroups"];
 };
 
 export async function fetchCurrencyAll({ providers, additionalCoinsFlag = false }: Props) {
-  if (isIntegrationTestEnv()) return Promise.resolve(fetchCurrencyAllMock);
+  if (isIntegrationTestEnv()) return Promise.resolve(flattenV5CurrenciesAll(fetchCurrencyAllMock));
 
   const url = new URL(`${getEnv("SWAP_API_BASE_V5")}/currencies/all`);
   url.searchParams.append("providers-whitelist", providers.join(","));
   url.searchParams.append("additional-coins-flag", additionalCoinsFlag.toString());
 
   try {
-    const { data } = await network<ResponseData>({
+    const { data } = await network<ResponseDataAll>({
       method: "GET",
       url: url.toString(),
       timeout: DEFAULT_SWAP_TIMEOUT_MS,
     });
-    return data;
+    return flattenV5CurrenciesAll(data);
   } catch (e: unknown) {
     if (axios.isAxiosError(e)) {
       if (e.code === "ECONNABORTED") {
