+++ libs/ledger-live-common/src/exchange/swap/api/v5/index.ts
@@ -1,1 +1,4 @@
 export { fetchCurrencyTo } from "./fetchCurrencyTo";
+export { fetchCurrencyFrom } from "./fetchCurrencyFrom";
+export { fetchCurrencyAll } from "./fetchCurrencyAll";
+export { fetchRates } from "./fetchRates";
