+++ libs/ledger-live-common/src/exchange/swap/hooks/index.ts
@@ -1,11 +1,9 @@
 export * from "./usePickDefaultAccount";
 export * from "./usePickDefaultCurrency";
 export * from "./useSelectableCurrencies";
-export * from "./useSwapProviders";
 export * from "./useSwapTransaction";
 export * from "./useUpdateMaxAmount";
-export * from "./v5/useProviderRates";
 export * from "./useFromState";
 export * from "./useToState";
 export * from "./usePageState";
-export * from "./v5/useFetchCurrencyTo";
+export * from "./v5";
